battleArea changes
