+++ app/src/main/java/at/aau/gloryweapons/siegeanddestroy3d/game/activities/GameTurnsActivity.java
@@ -317,7 +317,7 @@ public class GameTurnsActivity extends AppCompatActivity {
 
         // finds the weakest ship
         if (currentCheat == null){
-            currentCheat = actualBattleArea.findWeakestShip();
+            currentCheat = controller.findWeakestShip(actualBattleArea);
         }
 
         //Toast.makeText(this, "schummeln aktiv", Toast.LENGTH_SHORT);
